pyproject.toml
[project]
name = "fhe-mcp-server"
version = "0.1.0"
description = "FHE MCP Server"
readme = "README.md"
requires-python = ">=3.11.4"
dependencies = [
    "concrete-ml==1.9.0",
    "fastapi>=0.115.12",
    "mcp[cli]>=1.6.0",
    "numpy>=1.26.4",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "setuptools==75.3.0",
    "uvicorn>=0.34.2",
    "wheel>=0.45.1",
]
